third_party/WebKit/LayoutTests/external/wpt/html/webappapis/scripting/processing-model-2/compile-error-in-body-onerror-expected.txt

CONSOLE ERROR: line 19: Uncaught SyntaxError: Unexpected token )
CONSOLE ERROR: line 17: Uncaught SyntaxError: Unexpected token )
This is a testharness.js-based test.
PASS window.onerror - compile error in <body onerror> 
Harness: the test ran to completion.
